+++ src/js/options.js
@@ -18,6 +18,21 @@ $(document).ready(function() {
     getAccount();
 });
 
+$(function() {
+    $('a[href*=#]:not([href=#])').click(function() {
+        if (location.pathname.replace(/^\//, '') == this.pathname.replace(/^\//, '') && location.hostname == this.hostname) {
+            var target = $(this.hash);
+            target = target.length ? target : $('[id=' + this.hash.slice(1) + ']');
+            if (target.length) {
+                $('#contentarea').scrollTo(target, {
+                    duration: 'slow'
+                });
+                return false;
+            }
+        }
+    });
+});
+
 $("button#save").click(function() {
     var warningPercentage = $("input#warningPercentage").val();
     var warningDays = $("input#warningDays").val();
@@ -69,31 +84,40 @@ function api(url, callback) {
     });
 }
 
+function loginRequired() {
+    $('section.login-required').html('You have to be logged in to use this feature.');
+}
+
 function getAccount() {
     var url = "https://real-debrid.com/api/account.php?out=json";
     api(url, function(data) {
-        $('#username').html(data.username);
-        $('#email').html(data.email);
-        $('#points').html(data.points);
-        $('#expiration').html(data['expiration-txt']);
-        var html = '';
-        var limited = data.limited;
-        for (var i = 0; i < limited.length; i++) {
-            var hoster = limited[i];
-            var total = hoster.limit + hoster.additional_traffic;
-            var used = Math.round((hoster.downloaded / total) * 100);
-            html += '<div class="row">';
-            html += '<div class="key">' + hoster.name + '</div>';
-            html += '<div class="value">';
-            html += '<div class="progress">';
-            html += '<div class="progress-bar" role="progressbar" aria-valuenow="' + used + '" aria-valuemin="0" aria-valuemax="100" style="width: ' + used + '%;">';
-            html += used + '%';
-            html += '</div>';
-            html += '</div>';
-            html += '</div>';
-            html += '</div>';
+        if (data.error) {
+            loginRequired();
+        } else {
+            $('#username').html(data.username);
+            $('#email').html(data.email);
+            $('#points').html(data.points);
+            $('#expiration').html(data['expiration-txt']);
+            var html = '';
+            var limited = data.limited;
+            for (var i = 0; i < limited.length; i++) {
+                var hoster = limited[i];
+                var total = hoster.limit + hoster.additional_traffic;
+                var used = Math.round((hoster.downloaded / total) * 100);
+                html += '<div class="row">';
+                html += '<div class="key"><img src="' + hoster.image + '" class="emblem" //> ' + hoster.name + '</div>';
+                html += '<div class="value">';
+                html += '<div class="progress">';
+                html += '<div class="progress-bar" role="progressbar" aria-valuenow="' + used + '" aria-valuemin="0" aria-valuemax="100" style="width: ' + used + '%;">';
+                html += used + '%';
+                html += '</div>';
+                html += '</div>';
+                html += '</div>';
+                html += '</div>';
+            }
+            $('#hosters').append(html);
         }
-        $('#hosters').append(html);
+
     });
 
 }
@@ -101,17 +125,22 @@ function getAccount() {
 function getDownloads() {
     var url = "https://real-debrid.com/api/downloads.php?out=json";
     api(url, function(data) {
-        var downloads = data.downloads;
-        var html = '';
-        for (var i = 0; i < downloads.length; i++) {
-            html += '<tr>';
-            html += '<td><img src="' + downloads[i].hoster_image + '" //></td>';
-            html += '<td>' + downloads[i].name + '</td>';
-            html += '<td>' + downloads[i].generated_date + '</td>';
-            html += '<td><a href="' + downloads[i].original + '" target="_blank">Link</a></td>';
-            html += '<td><a href="' + downloads[i].link + '" target="_blank">Link</a></td>';
-            html += '</tr>';
+        if (data.error) {
+            loginRequired();
+        } else {
+            var downloads = data.downloads;
+            var html = '';
+            for (var i = 0; i < downloads.length; i++) {
+                html += '<tr>';
+                html += '<td><img src="' + downloads[i].hoster_image + '" //></td>';
+                html += '<td>' + downloads[i].name + '</td>';
+                html += '<td>' + downloads[i].generated_date + '</td>';
+                html += '<td><a href="' + downloads[i].original + '" target="_blank">Link</a></td>';
+                html += '<td><a href="' + downloads[i].link + '" target="_blank">Link</a></td>';
+                html += '</tr>';
+            }
+            $('#downloads tbody').append(html);
         }
-        $('#downloads tbody').append(html);
+
     });
 }