+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Features:
 
  - Recent downloads card
  - Account information card
+ - Limited hosters card
+ - Page action added to supported hosters (icon in URL bar)
 
 ## 1.6.x